refactor(crabtalk): move daemon transport wiring to transport.rs
