+++ crates/crabtalk/src/daemon/mod.rs
@@ -3,20 +3,28 @@
 use crate::{DaemonConfig, hooks, storage::FsStorage};
 use anyhow::Result;
 use crabllm_core::Provider;
-use futures_util::{StreamExt, pin_mut};
 use runtime::Runtime;
 use std::collections::HashMap;
 use std::{
     path::{Path, PathBuf},
     sync::Arc,
 };
-use tokio::sync::{Mutex, RwLock, broadcast, mpsc, oneshot};
-use wcore::protocol::{api::Server, message::ClientMessage};
+use tokio::sync::{Mutex, RwLock, broadcast, oneshot};
+pub use transport::{bridge_shutdown, setup_tcp};
 use {
     builder::{BuildProvider, DefaultProvider, build_default_provider},
     event::EventBus,
     host::DaemonEnv,
 };
+
+#[cfg(unix)]
+pub use transport::setup_socket;
+
+pub mod builder;
+pub mod event;
+pub mod hook;
+pub mod host;
+mod transport;
 
 /// Per-conversation working directory overrides.
 pub type ConversationCwds = Arc<Mutex<HashMap<u64, PathBuf>>>;
@@ -24,10 +32,8 @@ pub type ConversationCwds = Arc<Mutex<HashMap<u64, PathBuf>>>;
 /// Pending ask_user oneshots (shared with AskUserHook and protocol layer).
 pub type PendingAsks = Arc<Mutex<HashMap<u64, oneshot::Sender<String>>>>;
 
-pub mod builder;
-pub mod event;
-pub mod hook;
-pub mod host;
+/// Shared runtime handle.
+pub type SharedRuntime<P> = Arc<RwLock<Arc<Runtime<DaemonCfg<P>>>>>;
 
 /// Config binding for a node.
 pub struct DaemonCfg<P: Provider + 'static = DefaultProvider> {
@@ -39,9 +45,6 @@ impl<P: Provider + 'static> runtime::Config for DaemonCfg<P> {
     type Provider = P;
     type Env = DaemonEnv;
 }
-
-/// Shared runtime handle.
-pub type SharedRuntime<P> = Arc<RwLock<Arc<Runtime<DaemonCfg<P>>>>>;
 
 /// Shared daemon state.
 pub struct Daemon<P: Provider + 'static = DefaultProvider> {
@@ -112,77 +115,4 @@ impl<P: Provider + 'static> DaemonHandle<P> {
         let _ = self.shutdown_tx.send(());
         Ok(())
     }
-}
-
-// ── Transport setup helpers ──────────────────────────────────────────
-
-fn dispatch_callback<P: Provider + 'static>(
-    daemon: Daemon<P>,
-) -> impl Fn(ClientMessage, mpsc::Sender<wcore::protocol::message::ServerMessage>) + Clone + Send + 'static
-{
-    move |msg, reply| {
-        let daemon = daemon.clone();
-        tokio::spawn(async move {
-            let stream = daemon.dispatch(msg);
-            pin_mut!(stream);
-            while let Some(server_msg) = stream.next().await {
-                if reply.send(server_msg).await.is_err() {
-                    break;
-                }
-            }
-        });
-    }
-}
-
-#[cfg(unix)]
-pub fn setup_socket<P: Provider + 'static>(
-    daemon: Daemon<P>,
-    shutdown_tx: &broadcast::Sender<()>,
-) -> Result<(&'static Path, tokio::task::JoinHandle<()>)> {
-    let resolved_path: &'static Path = &wcore::paths::SOCKET_PATH;
-    if let Some(parent) = resolved_path.parent() {
-        std::fs::create_dir_all(parent)?;
-    }
-    if resolved_path.exists() {
-        std::fs::remove_file(resolved_path)?;
-    }
-
-    let listener = tokio::net::UnixListener::bind(resolved_path)?;
-    tracing::info!("daemon listening on {}", resolved_path.display());
-
-    let socket_shutdown = bridge_shutdown(shutdown_tx.subscribe());
-    let join = tokio::spawn(transport::uds::accept_loop(
-        listener,
-        dispatch_callback(daemon),
-        socket_shutdown,
-    ));
-
-    Ok((resolved_path, join))
-}
-
-pub fn setup_tcp<P: Provider + 'static>(
-    daemon: Daemon<P>,
-    shutdown_tx: &broadcast::Sender<()>,
-) -> Result<(tokio::task::JoinHandle<()>, u16)> {
-    let (std_listener, addr) = transport::tcp::bind()?;
-    let listener = tokio::net::TcpListener::from_std(std_listener)?;
-    tracing::info!("daemon listening on tcp://{addr}");
-
-    let tcp_shutdown = bridge_shutdown(shutdown_tx.subscribe());
-    let join = tokio::spawn(transport::tcp::accept_loop(
-        listener,
-        dispatch_callback(daemon),
-        tcp_shutdown,
-    ));
-
-    Ok((join, addr.port()))
-}
-
-pub fn bridge_shutdown(mut rx: broadcast::Receiver<()>) -> oneshot::Receiver<()> {
-    let (otx, orx) = oneshot::channel();
-    tokio::spawn(async move {
-        let _ = rx.recv().await;
-        let _ = otx.send(());
-    });
-    orx
 }
